backend/requirements.txt
anthropic==0.40.0
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==14.1
python-dotenv==1.0.1
aiosqlite==0.20.0
numpy==2.1.3
scipy==1.14.1
pydantic==2.10.1
httpx==0.27.2
python-multipart==0.0.18
aiofiles==24.1.0
